Setup process working for conda env creation and downloading squeezenet pretrained pytorch weights

+++ setup/setup.sh
@@ -25,38 +25,57 @@ function get_platform() {
     set -e
 }
 
+function check_conda_env_exists() {
+    source activate labelar_demo
+    if [ $? -eq 0 ]; then
+        echo "EXISTS"
+    else
+        echo "DOES NOT EXIST"
+    fi
+}
+
+# Need this on some systems to enable conda from bash:
+eval "$(conda shell.bash hook)"
+check_conda_env_exists
+# exit 0
+
+# Assume we are in ./setup/ when script is called
+
 # Detect platform:
 platform="cpu"
 get_platform $platform
 
 # Delete and recreate conda env:
-eval "$(conda shell.bash hook)"
-conda deactivate
-conda env remove -n labelar_demo
-conda env create -f environment-$platform.yml
-if [[ "${platform}" == "cpu" ]]; then
-    mv environment-cpu.yml.bak environment-cpu.yml
+REBUILD_ENV=true
+if [[ "$REBUILD_ENV" = true ]]; then
+    echo "(Re)-building labelar_demo conda env ${REBUILD_ENV}"
+    conda deactivate
+    conda env remove -n labelar_demo
+    conda env create -f environment-$platform.yml
+    if [[ "${platform}" == "cpu" ]]; then
+        if [[ -f environment-cpu.yml.bak ]]; then
+            mv environment-cpu.yml.bak environment-cpu.yml
+        fi
+    fi
 fi
+
 
 # Enable using the h4d_env conda environment from jupyter notebooks:
 conda activate labelar_demo
 python -m ipykernel install --user --name h4d_env --display-name "labelar_demo"
 # Install pycocotools:
-cd cocoapi/PythonAPI
-python3 setup.py build_ext install
-cd ../..
-# Install def_conv:
-cd vendor/def_conv
+cd ../vendor/cocoapi/PythonAPI
 python3 setup.py build_ext install
 cd -
 
 
-if [[ "${platform}" == "gpu" ]]; then
-    # Do GPU specific setup here, e.g., CenterNet, DeformConvVxx.
-fi
+# if [[ "${platform}" == "gpu" ]]; then
+#     # Do GPU specific setup here, e.g., CenterNet, DeformConvVxx.
+# fi
 
 
 # Download weights:
-mkdir -p ./training/weights/pretrained/
-cd ./training/weights/pretrained/
+mkdir -p ../training/weights/pretrained/
+cd ../training/weights/pretrained/
 wget https://download.pytorch.org/models/squeezenet1_1-f364aa15.pth
+cd -
